LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-rectangles-to-the-canvas/2d.clearRect.transform-expected.txt

2d.clearRect.transform
clearRect is affected by transforms
Actual output:
Expected output:


PASS clearRect is affected by transforms
